# lab2/s1/src/main/resources/openapi.json
{"openapi":"3.0.1","info":{"title":"Team manager API","version":"1.0"},"servers":[{"url":"https://rwfsh39.ru/s1/api","description":"First Web Server"},{"url":"https://rwfsh39.ru/s2/api","description":"Second Web Server"}],"paths":{"/v1/human-being/{id}":{"get":{"tags":["human-being"],"summary":"Get Human Being","description":"Get Human Being by id from collection","operationId":"getHumanBeingById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Human Being instance retrieved","content":{"application/xml":{"schema":{"$ref":"#/components/schemas/HumanBeingDto"}}}},"400":{"description":"Operations parameters are invalid"}}},"put":{"tags":["human-being"],"summary":"Modify Human Being","description":"Modify existing Human Being instance in collection","operationId":"modifyHumanBeing","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/xml":{"schema":{"$ref":"#/components/schemas/modifyHumanBeingRequest"}}},"required":true},"responses":{"200":{"description":"Human Being has been modified","content":{"application/xml":{"schema":{"$ref":"#/components/schemas/HumanBeingDto"}}}},"400":{"description":"Operations parameters are invalid"},"415":{"description":"Content-Type header must be specified as XML in request"}}}},"/v1/human-being":{"get":{"tags":["human-being"],"summary":"Get all human-being","description":"Get all human-being instances in collection","operationId":"getAllHumanBeing","parameters":[{"name":"idGte","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"idLte","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"nameIn","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"}}},{"name":"coordinateXGte","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"coordinateXLte","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"coordinateYGte","in":"query","required":false,"schema":{"type":"number","format":"double"}},{"name":"coordinateYLte","in":"query","required":false,"schema":{"type":"number","format":"double"}},{"name":"creationDateGte","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"creationDateLte","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"realHero","in":"query","required":false,"schema":{"type":"boolean"}},{"name":"hasToothpick","in":"query","required":false,"schema":{"type":"boolean"}},{"name":"impactSpeedGte","in":"query","required":false,"schema":{"type":"number","format":"float"}},{"name":"impactSpeedLte","in":"query","required":false,"schema":{"type":"number","format":"float"}},{"name":"minutesOfWaitingGte","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"minutesOfWaitingLte","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"coolCar","in":"query","required":false,"schema":{"type":"boolean"}},{"name":"moodIn","in":"query","required":false,"schema":{"type":"array","items":{"title":"Mood enum","type":"string","enum":["SORROW","APATHY","CALM","RAGE","FRENZY"]}}},{"name":"weaponTypeIn","in":"query","required":false,"schema":{"type":"array","items":{"title":"Weapon type enum","type":"string","enum":["HAMMER","AXE","PISTOL","KNIFE","BAT"]}}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sortDirections","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"}}},{"name":"sortFields","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"Human Being results from datasource","content":{"application/xml":{"schema":{"$ref":"#/components/schemas/ListHumanBeingDto"}}}},"400":{"description":"Operations parameters are invalid"}}},"post":{"tags":["human-being"],"summary":"Create new Human Being","description":"Create new Human Being in collection","operationId":"createHumanBeing","requestBody":{"content":{"application/xml":{"schema":{"$ref":"#/components/schemas/createHumanBeingRequest"}}},"required":true},"responses":{"200":{"description":"Human Being has been created","content":{"application/xml":{"schema":{"$ref":"#/components/schemas/HumanBeingDto"}}}},"400":{"description":"Operations parameters are invalid"},"415":{"description":"Content-Type header must be specified as XML in request"}}}},"/v1/human-being/stats":{"get":{"tags":["human-being"],"summary":"Get human-being statistic","description":"Get human-being statistic","operationId":"getHumanStats","parameters":[{"name":"field","in":"query","required":false,"schema":{"title":"Statistic field DTO","type":"string","enum":["IMPACT_SPEED","MINUTES_OF_WAITING"]}},{"name":"operation","in":"query","required":false,"schema":{"title":"Statistic operation DTO","type":"string","enum":["MEAN","MIN","MAX"]}}],"responses":{"200":{"description":"Human Being results from datasource","content":{"application/xml":{"schema":{"$ref":"#/components/schemas/StatisticResponseDto"}}}},"400":{"description":"Operations parameters are invalid"}}}},"/v1/human-being/":{"delete":{"tags":["human-being"],"summary":"Delete human being with given params and limit","description":"Delete human being by given car or impact speed, with limit","operationId":"deleteByParams","parameters":[{"name":"carCool","in":"query","required":false,"schema":{"type":"boolean"}},{"name":"impactSpeed","in":"query","required":false,"schema":{"type":"number","format":"float"}},{"name":"limit","in":"query","required":false,"schema":{"minimum":0,"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Result of delete operation"},"400":{"description":"Operations parameters are invalid"}}}}},"components":{"schemas":{"CarDto":{"type":"object","properties":{"cool":{"type":"boolean","description":"Is car really cool?","xml":{"attribute":true}}},"description":"Human Car"},"CoordinatesDto":{"required":["y"],"type":"object","properties":{"x":{"type":"integer","description":"X coordinate","format":"int32"},"y":{"type":"number","description":"Y coordinate","format":"float"}},"description":"Coordinates"},"HumanBeingDto":{"title":"HumanBeing","required":["coordinates","creationDate","hasToothpick","id","minutesOfWaiting","name","realHero"],"type":"object","properties":{"id":{"type":"integer","description":"Unique ID generated on server","format":"int32"},"name":{"type":"string","description":"Name of Human Being"},"coordinates":{"$ref":"#/components/schemas/CoordinatesDto"},"creationDate":{"type":"string","description":"When object was created","format":"date-time"},"realHero":{"type":"boolean","description":"Is human real hero?","xml":{"attribute":true}},"hasToothpick":{"type":"boolean","description":"Does human has toothpick?","xml":{"attribute":true}},"impactSpeed":{"type":"number","description":"Impact speed","format":"float"},"minutesOfWaiting":{"type":"integer","description":"Minutes of waiting","format":"int64"},"weaponType":{"title":"Weapon type enum","type":"string","description":"Weapon Type","enum":["HAMMER","AXE","PISTOL","KNIFE","BAT"]},"mood":{"title":"Mood enum","type":"string","description":"Mood","enum":["SORROW","APATHY","CALM","RAGE","FRENZY"]},"car":{"$ref":"#/components/schemas/CarDto"}},"description":"Main structural unit. Describes a person. Purpose of this object is to be a part of some collection."},"modifyHumanBeingRequest":{"required":["car","coordinates","mood","weaponType"],"type":"object","properties":{"name":{"maxLength":2147483647,"minLength":1,"type":"string","description":"Name of Human Being"},"coordinates":{"$ref":"#/components/schemas/CoordinatesDto"},"realHero":{"type":"boolean","description":"Is real hero?","xml":{"attribute":true}},"hasToothpick":{"type":"boolean","description":"Does hero have a toothpick?","xml":{"attribute":true}},"impactSpeed":{"maximum":676,"minimum":0,"type":"number","description":"Impact Speed","format":"float"},"minutesOfWaiting":{"minimum":0,"type":"integer","description":"Minutes of waiting","format":"int64"},"car":{"$ref":"#/components/schemas/CarDto"},"mood":{"title":"Mood enum","type":"string","description":"Mood type","enum":["SORROW","APATHY","CALM","RAGE","FRENZY"]},"weaponType":{"title":"Weapon type enum","type":"string","description":"Weapon Type","enum":["HAMMER","AXE","PISTOL","KNIFE","BAT"]}}},"createHumanBeingRequest":{"required":["car","coordinates","hasToothpick","impactSpeed","minutesOfWaiting","mood","name","realHero","weaponType"],"type":"object","properties":{"name":{"maxLength":2147483647,"minLength":1,"type":"string","description":"Name of Human Being"},"coordinates":{"$ref":"#/components/schemas/CoordinatesDto"},"realHero":{"type":"boolean","description":"Is real hero?","xml":{"attribute":true}},"hasToothpick":{"type":"boolean","description":"Does hero have a toothpick?","xml":{"attribute":true}},"impactSpeed":{"maximum":676,"minimum":0,"type":"number","description":"Impact Speed","format":"float"},"minutesOfWaiting":{"minimum":0,"type":"integer","description":"Minutes of waiting","format":"int64"},"car":{"$ref":"#/components/schemas/CarDto"},"mood":{"title":"Mood enum","type":"string","description":"Mood type","enum":["SORROW","APATHY","CALM","RAGE","FRENZY"]},"weaponType":{"title":"Weapon type enum","type":"string","description":"Weapon Type","enum":["HAMMER","AXE","PISTOL","KNIFE","BAT"]}}},"ListHumanBeingDto":{"required":["humanBeing"],"type":"object","properties":{"humanBeing":{"type":"array","description":"Result human being","items":{"$ref":"#/components/schemas/HumanBeingDto"}}},"description":"List of human being"},"StatisticResponseDto":{"title":"statisticResult","type":"object","properties":{"result":{"type":"number","description":"Statistic Result","format":"double"}},"description":"Result of statistic request"}}}}